Migrate iris.experimental.animate to iris.plot.
